feat(auth): move login page to app router
